refactoring-menu-after-new-feature
componentization and styles
