+++ src/components/Menu/Pizza/AddToCart/style.scss
@@ -1,0 +1,41 @@
+@import '../../../../style/abstract/variables.scss';
+@mixin iconCart {
+    font-size: 2rem;
+    font-weight: bold;
+    text-align: center;
+    justify-items: center;
+    border-radius: 50%;
+    padding: 5px;   
+    background: $tertiary-color;
+    transform: scale(.9);
+    transition: 0.3s;
+}
+
+.menu__item--addCart{
+    @include iconCart();
+    &:hover{
+        background-color: $secundary-color;
+        transform: scale(1);
+        transition: 0.3s;
+        cursor: pointer;
+    }   
+}
+.menu__item--cartCheck{
+    background-color: $secundary-color;
+    transform: scale(1);
+    color: white;
+}
+.menu__item--cart{
+    display: none;
+}
+.menu__item--remove{
+    font-size: 2rem;
+    font-weight: bold;
+    text-align: center;
+    justify-items: center;
+    border-radius: 50%;
+    padding: 5px;   
+    background: rgb(248, 162, 162);
+    transform: scale(.9);
+    transition: 0.3s;
+}
