+++ src/components/Menu/Pizza/Menu/style.scss
@@ -1,15 +1,5 @@
 @import "../../../../style/abstract/variables.scss";
-@mixin iconCart {
-    font-size: 2rem;
-    font-weight: bold;
-    text-align: center;
-    justify-items: center;
-    border-radius: 50%;
-    padding: 5px;   
-    background: $tertiary-color;
-    transform: scale(.9);
-    transition: 0.3s;
-}
+
 @mixin promotionDay {
     display: flex;
     justify-content: center;
@@ -58,26 +48,7 @@
             position: relative;
             box-sizing: border-box;
             overflow: hidden;
-            .tag-cart{
-                @include promotionDay();
-                background-color: $secundary-color;
-                color: white;
-            }
-            .tag-promotion{
-                @include promotionDay();
-                background-color: $primary-color;
-                color: white;
-            }
-            &:hover{
-                z-index: 9999999;
-                border-radius: 10px;
-                box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
-                transition: .3s;
-                @media (max-width: 760px) {
-                    transform: scale(1);
-                    box-shadow: none;
-                }
-            }    
+   
             .menu__item--image{
                 width: 100px;
                 height: 100px;
@@ -104,81 +75,7 @@
                 .menu__item--description{
                     color: rgb(80, 80, 80);
                 }
-                .menu__item--content{
-                    display: flex;
-                    justify-content: flex-end;
-                    align-items: center;
-                    gap: 20px;
-                    margin-top: 10px;
-                    @media (max-width: 760px) {
-                        gap: 10px;
-                    }
-                    .menu__item--content--price{
-                        display: flex;
-                        gap: 20px;
-                        @media (max-width: 760px) {
-                            flex-direction: column;
-                            gap: 0;
-                        }
-                        .menu__item--price{
-                            font-weight: bold;
-                            font-size: 20px;
-                            &::before{
-                                content: 'R$';
-                                margin-right: 5px;
-                            }
-                        }
-                        .menu__item__promotion--price-show{
-                            display: block;
-                            color: $secundary-color;
-                            &::before{
-                                content: 'R$';
-                                margin-right: 5px;
-                            }
-                        }
-                        .menu__item__promotion--price-hide{
-                            display: none;
-                        }
-                        .menu__item--old-price{
-                            text-decoration: line-through;
-                            font-size: 16px;
-                            color: red;
-                            &::before{
-                                content: 'R$';
-                                margin-right: 5px;
-                            }
-                    }
-                    
-                    }
-                    .menu__item--addCart{
-                        @include iconCart();
-                        &:hover{
-                            background-color: $secundary-color;
-                            transform: scale(1);
-                            transition: 0.3s;
-                            cursor: pointer;
-                        }   
-                    }
-                    .menu__item--cartCheck{
-                        background-color: $secundary-color;
-                        transform: scale(1);
-                        color: white;
-                    }
-                    .menu__item--cart{
-                        display: none;
-                    }
-                    .menu__item--remove{
-                        font-size: 2rem;
-                        font-weight: bold;
-                        text-align: center;
-                        justify-items: center;
-                        border-radius: 50%;
-                        padding: 5px;   
-                        background: rgb(248, 162, 162);
-                        transform: scale(.9);
-                        transition: 0.3s;
-                    } 
-                }  
+
             }        
         } 
     }
