+++ src/components/Menu/Pizza/ProductPrice/style.scss
@@ -1,0 +1,60 @@
+@import '../../../../style/abstract/variables.scss';
+@mixin iconCart {
+    font-size: 2rem;
+    font-weight: bold;
+    text-align: center;
+    justify-items: center;
+    border-radius: 50%;
+    padding: 5px;   
+    background: $tertiary-color;
+    transform: scale(.9);
+    transition: 0.3s;
+}
+.menu__item--content{
+    display: flex;
+    justify-content: flex-end;
+    align-items: center;
+    gap: 20px;
+    margin-top: 10px;
+    @media (max-width: 760px) {
+        gap: 10px;
+    }
+    .menu__item--content--price{
+        display: flex;
+        gap: 20px;
+        @media (max-width: 760px) {
+            flex-direction: column;
+            gap: 0;
+        }
+        .menu__item--price{
+            font-weight: bold;
+            font-size: 20px;
+            &::before{
+                content: 'R$';
+                margin-right: 5px;
+            }
+        }
+        .menu__item__promotion--price-show{
+            display: block;
+            color: $secundary-color;
+            &::before{
+                content: 'R$';
+                margin-right: 5px;
+            }
+        }
+        .menu__item__promotion--price-hide{
+            display: none;
+        }
+        .menu__item--old-price{
+            text-decoration: line-through;
+            font-size: 16px;
+            color: red;
+            &::before{
+                content: 'R$';
+                margin-right: 5px;
+            }
+    }
+    
+    }
+
+}
